// server/src/features/icebreaker-messages-request/errors/enum.ts
export enum FeatureError {
  BAD_REQUEST = 'bad_request',
  SENDER_PROFILE = 'sender_profile',
  RECEIVER_PROFILE = 'receiver_profile',
  SENDER_POSTS = 'sender_posts',
  RECEIVER_POSTS = 'receiver_posts',
  SENDER_COMMENTS = 'sender_comments',
  RECEIVER_REACTIONS = 'receiver_reactions',
  MESSAGES_TEMPLATE = 'messages_template',
  AI_RESPONSE = 'ai_response',
  LINKEDIN_API_RATE_LIMIT = 'linkedin_api_rate_limit',
  UNEXPECTED_ERROR = 'unexpected_error',
}